admin dashboard make dynamic done
