+++ src/services/AdminDashboard/index.ts
@@ -1,0 +1,25 @@
+"use server";
+import { cookies } from "next/headers";
+
+export async function getAdminDashboardInfo() {
+  try {
+    const response = await fetch(
+      `${process.env.NEXT_PUBLIC_BASE_API}/payment/admin-dashboard-info`,
+      {
+        method: "GET",
+        headers: {
+          Authorization: ` ${(await cookies()).get("accessToken")!.value}`,
+        },
+      }
+    );
+
+    const data = await response.json();
+    return { data:data.data };
+  } catch (error) {
+    // console.error("Error fetching review:", error);
+    return {
+      data: null,
+      error: error instanceof Error ? error.message : "Unknown error",
+    };
+  }
+}
